requirements.txt
astroid==1.6.5
autopep8==1.4.3
backports.functools-lru-cache==1.5
certifi==2018.11.29
chardet==3.0.4
configparser==3.5.0
coreapi==2.3.3
coreschema==0.0.4
Django==1.11
django-extensions==2.1.4
django-filter==2.0.0
djangorestframework==3.8.0
enum34==1.1.6
futures==3.2.0
idna==2.8
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
Markdown==3.0.1
MarkupSafe==1.1.0
mccabe==0.6.1
pbr==5.1.1
psycopg2-binary==2.7.6.1
pycodestyle==2.4.0
pylint==1.9.4
pytz==2018.9
requests==2.21.0
singledispatch==3.4.0.3
six==1.12.0
stevedore==1.30.0
typing==3.6.6
uritemplate==3.0.0
urllib3==1.24.1
virtualenv==16.2.0
virtualenv-clone==0.4.0
virtualenvwrapper==4.8.2
wrapt==1.11.0
